requirements.txt
langchain
langchain-google-genai
langgraph
python-dotenv
twilio
